pyproject.toml
[project]
name = "test-a2a-agent"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=0.3.22",
]
